refactor: move module loader to its own file
This is pure refactoring with no functional changes.
