+++ runtime/lib.rs
@@ -3,6 +3,8 @@ pub use deno_core;
 pub mod runtime;
 pub use runtime::*;
 
+mod module_loader;
+
 mod vendored;
 pub use vendored::colors;
 pub use vendored::fmt_errors;
